+++ src/main/java/com/v_ia_backend/kipa/dto/response/MovementTableResponse.java
@@ -4,6 +4,7 @@ import com.v_ia_backend.kipa.entity.*;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
@@ -11,12 +12,12 @@ import java.util.List;
 public class MovementTableResponse {
     private Long id;
     private HigherAccounts higherAccountId;
-    private Long debit;
-    private Long credit;
-    private Long balance;
+    private BigDecimal debit;
+    private BigDecimal credit;
+    private BigDecimal balance;
     private List<MovementListResponse> movementListResponse;
 
-    public MovementTableResponse(Movements movements, Long debit, Long credit, Long balance, List<MovementListResponse> movementListResponse) {
+    public MovementTableResponse(Movements movements, BigDecimal debit, BigDecimal credit, BigDecimal balance, List<MovementListResponse> movementListResponse) {
         this.id = movements.getId();
         this.higherAccountId = movements.getHigherAccountId();
         this.debit = debit;
